requirements.txt
flask==3.1.0
anthropic==0.52.0
python-dotenv==1.0.1
gunicorn==23.0.0
